fix(scheduler): add duration executor
